code clear (fix) hometask YA
Yevhenii Andrushchenko
