+++ src/main/java/core/basesyntax/HelloWorld.java
@@ -1,0 +1,43 @@
+package core.basesyntax;
+
+import core.basesyntax.model.Cat;
+import core.basesyntax.model.Dog;
+
+import java.time.LocalDate;
+
+public class HelloWorld {
+
+    public static void main(String[] args) {
+        Dog dog = new Dog();
+        Cat cat = new Cat();
+    }
+
+    private String getFromOneToTen() {
+        StringBuilder result = new StringBuilder();
+
+        for (int i = 0; i < 10; i++) {
+            if (i % 2 == 0) {
+                result.append(i).append(" ");
+            }
+        }
+        return result.toString();
+    }
+
+    private void sayHello(String helloMessage) {
+        System.out.println("Hello, Mate! I am happy to see you! This class looks weird! "
+                + "Your task is to fix the checkstyle in this file."
+                + " Your goal is to run `mvn clean package` and receive `BUILD SUCCESS` message");
+        System.out.println(helloMessage);
+    }
+
+    private void initializeVariables() {
+        String a = "Hello mates!";
+        int variable = LocalDate.now().getYear();
+        int age = 25;
+        int size = 42;
+        System.out.println(a);
+        System.out.println("It is " + variable + "year");
+        System.out.println("I'm " + age);
+        System.out.println("The size is " + size);
+    }
+}
