--- src/main/java/core/basesyntax/Model/Dog.java
@@ -1,4 +1,0 @@
-package core.basesyntax.Model;
-
-public class Dog {
-}
