+++ src/main/java/core/basesyntax/model/Cat.java
@@ -1,4 +1,4 @@
-package core.basesyntax.Model;
+package core.basesyntax.model;
 
 public class Cat {
     private int age;
